Add AttributeFilter
Added an AttributeFilter class that is subclassed by the different types
of attributeFilter. Added an AttributeMapping as part of the Mapping.
AttributeFilter are converted to CQL by first converting them to
modifiers using the toModifier function which takes the AttributeMapping
as input. Afterwards the Modifiers are treated the same as the existing
Modifiers. Added new Modifiers for RangeCriterion and NumericCriterion.