Add Patient Criteria a Coding Value Type
The ValueSetCriterion can now handle criteria on the Patient. The
Patient has to be access differently to other Resources, because the
query runs in the context of one single Patient. So we don't need to use
a retrieve expression to access the patient. Instead we can simply use
the Patient identifier. In order to have the patient identifier
available, we need to declare the query context Patient explicitly.

Also in ValueSetCriterion it's now possible to compare against values of
type Coding instead of only CodeableConcept before.